day08/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-myimg==0.1
gunicorn==20.0.4
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
